feature: get subject by department id
