+++ apps/lxp/modules/training-information/content/api.ts
@@ -1,4 +1,4 @@
-import api from '../../services/api';
+import api from '../../../services/api';
 import { TSubjectResponse } from './types';
 
 export const getSubjectByDeptId = async (
