+++ apps/lxp/modules/training-information/content/detail/detail.tsx
