+++ apps/lxp/modules/training-information/content/detail/index.ts
