+++ apps/lxp/modules/training-information/content/detail/types.d.ts
