+++ apps/lxp/modules/training-information/content/header/index.ts
