+++ apps/lxp/modules/training-information/content/header/store.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const TrainingConfirmPopup = atom({
+  key: 'confirm-training',
+  default: false,
+});
