+++ apps/lxp/modules/training-information/content/header/types.d.ts
@@ -1,0 +1,6 @@
+export type TSubjectHeaderProps = {
+  name: string;
+  category?: string;
+  point?: number;
+  batch?: number;
+};
