+++ apps/lxp/modules/training-information/content/hook.ts
