+++ apps/lxp/modules/training-information/content/index.ts
@@ -1,0 +1,1 @@
+export * from './section';
