+++ apps/lxp/modules/training-information/content/store.ts
