+++ apps/lxp/modules/training-information/content/table/index.ts
