+++ apps/lxp/modules/training-information/content/table/store.ts
