+++ apps/lxp/modules/training-information/content/table/types.d.ts
@@ -1,0 +1,5 @@
+import { ISubjectDepartmentItem } from '../types';
+
+export type TsubjectTableProps = {
+  subjectData: ISubjectDepartmentItem[];
+};
