--- apps/lxp/modules/training-information/main/index.ts
@@ -1,1 +1,0 @@
-export * from './main';
