--- apps/lxp/services/types.d.ts
@@ -1,7 +1,0 @@
-export interface ICommonMetaResponse<T> {
-  code?: number;
-  status?: string;
-  message?: string;
-  data?: Array<T>;
-}
-export type TMetaErrorResponse = AxiosError<TMetaItem>;
